feature search screen refactoring
